+++ linkerd/admin/src/main/scala/io/buoyant/linkerd/admin/LinkerdAdmin.scala
@@ -8,8 +8,9 @@ import com.twitter.server.Admin.Path
 import com.twitter.server.handler.{SummaryHandler => TSummaryHandler, _}
 import com.twitter.server.view.{IndexView, TextBlockView}
 import io.buoyant.linkerd.Linker
+import io.buoyant.linkerd.Linker.LinkerConfig
 
-class LinkerdAdmin(app: App, linker: Linker) {
+class LinkerdAdmin(app: App, linker: Linker, config: LinkerConfig) {
 
   private[this] val log = Logger()
 
@@ -48,6 +49,7 @@ class LinkerdAdmin(app: App, linker: Linker) {
     }
 
   private[this] def linkerdAdminRoutes: Seq[(String, Service[Request, Response])] = Seq(
+
     "/" -> new SummaryHandler(linker),
     "/files/" -> (StaticFilter andThen ResourceHandler.fromDirectoryOrJar(
       baseRequestPath = "/files/",
@@ -57,7 +59,8 @@ class LinkerdAdmin(app: App, linker: Linker) {
     "/delegator" -> DelegateHandler.ui(linker),
     "/delegator.json" -> DelegateHandler.api(linker),
     "/routers.json" -> new RouterHandler(linker),
-    "/metrics" -> MetricsHandler
+    "/metrics" -> MetricsHandler,
+    "/config.json" -> new ConfigHandler(config, Linker.configInitializers)
   )
 
   private[this] def metricsRoutes: Seq[(String, Service[Request, Response])] = Seq(