+++ linkerd/config/src/main/scala/io/buoyant/linkerd/config/Parser.scala
@@ -1,8 +1,10 @@
 package io.buoyant.linkerd.config
 
+import com.fasterxml.jackson.annotation.JsonInclude.Include
 import com.fasterxml.jackson.core.JsonFactory
 import com.fasterxml.jackson.databind.deser.std.StdDeserializer
 import com.fasterxml.jackson.databind.module.SimpleModule
+import com.fasterxml.jackson.databind.ser.std.StdSerializer
 import com.fasterxml.jackson.databind.{DeserializationContext, ObjectMapper}
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
@@ -10,7 +12,11 @@ import com.fasterxml.jackson.module.scala.experimental.ScalaObjectMapper
 import com.twitter.finagle.util.LoadService
 import scala.reflect.{ClassTag, classTag}
 
-abstract class ConfigDeserializer[T: ClassTag] extends StdDeserializer[T](Parser.jClass[T]) {
+private[config] trait Registration {
+  def register(module: SimpleModule): SimpleModule
+}
+
+abstract class ConfigDeserializer[T: ClassTag] extends StdDeserializer[T](Parser.jClass[T]) with Registration {
   def register(module: SimpleModule): SimpleModule = module.addDeserializer(Parser.jClass[T], this)
 
   protected def catchMappingException(ctxt: DeserializationContext)(t: => T): T =
@@ -18,6 +24,10 @@ abstract class ConfigDeserializer[T: ClassTag] extends StdDeserializer[T](Parser
       case arg: IllegalArgumentException =>
         throw ctxt.mappingException(arg.getMessage)
     }
+}
+
+abstract class ConfigSerializer[T: ClassTag] extends StdSerializer[T](Parser.jClass[T]) with Registration {
+  def register(module: SimpleModule): SimpleModule = module.addSerializer(Parser.jClass[T], this)
 }
 
 object Parser {
@@ -31,9 +41,23 @@ object Parser {
    * be Json. We expose this publicly for testing purposes (to allow easy
    * parsing of config subtrees) at the moment.
    */
-  def objectMapper(config: String): ObjectMapper with ScalaObjectMapper = {
+  def objectMapper(
+    config: String,
+    configInitializers: Seq[ConfigInitializer]
+  ): ObjectMapper with ScalaObjectMapper = {
     val factory = if (peekJsonObject(config)) new JsonFactory() else new YAMLFactory()
-    val customTypes = LoadService[ConfigDeserializer[_]]
+    objectMapper(factory, configInitializers)
+  }
+
+  def jsonObjectMapper(
+    configInitializers: Seq[ConfigInitializer]
+  ): ObjectMapper with ScalaObjectMapper = objectMapper(new JsonFactory(), configInitializers)
+
+  private[this] def objectMapper(
+    factory: JsonFactory,
+    configInitializers: Seq[ConfigInitializer]
+  ): ObjectMapper with ScalaObjectMapper = {
+    val customTypes = (LoadService[ConfigDeserializer[_]] ++ LoadService[ConfigSerializer[_]])
       .foldLeft(new SimpleModule("linkerd custom types")) { (module, d) =>
         d.register(module)
       }
@@ -41,6 +65,15 @@ object Parser {
     val mapper = new ObjectMapper(factory) with ScalaObjectMapper
     mapper.registerModule(DefaultScalaModule)
     mapper.registerModule(customTypes)
+    mapper.setSerializationInclusion(Include.NON_NULL)
+
+    // Subtypes must not conflict
+    configInitializers.groupBy(_.configId).collect {
+      case (id, cis) if cis.size > 1 =>
+        throw ConflictingSubtypes(cis(0).namedType, cis(1).namedType)
+    }
+    for (ci <- configInitializers) ci.registerSubtypes(mapper)
+
     mapper
   }
 }