+++ linkerd/config/src/main/scala/io/buoyant/linkerd/config/types/PortDeserializer.scala
@@ -1,8 +1,8 @@
 package io.buoyant.linkerd.config.types
 
-import com.fasterxml.jackson.core.JsonParser
-import com.fasterxml.jackson.databind.DeserializationContext
-import io.buoyant.linkerd.config.ConfigDeserializer
+import com.fasterxml.jackson.core.{JsonGenerator, JsonParser}
+import com.fasterxml.jackson.databind.{SerializerProvider, DeserializationContext}
+import io.buoyant.linkerd.config.{ConfigSerializer, ConfigDeserializer}
 
 case class Port(port: Int) {
   val MinValue = 0
@@ -15,4 +15,12 @@ class PortDeserializer extends ConfigDeserializer[Port] {
     catchMappingException(ctxt) {
       Port(_parseInteger(jp, ctxt))
     }
+}
+
+class PortSerializer extends ConfigSerializer[Port] {
+  override def serialize(
+    value: Port,
+    jgen: JsonGenerator,
+    provider: SerializerProvider
+  ): Unit = jgen.writeNumber(value.port)
 }