+++ linkerd/core/src/main/scala/io/buoyant/linkerd/NamerInitializer.scala
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility
 import com.fasterxml.jackson.annotation.{JsonAutoDetect, JsonIgnore, JsonProperty, JsonTypeInfo}
 import com.twitter.finagle.{Path, Namer => FinagleNamer, Stack}
 import com.twitter.finagle.naming.NameInterpreter
+import io.buoyant.linkerd.config.ConfigInitializer
 
 sealed trait NamingFactory
 
@@ -31,7 +32,7 @@ object NamingFactory {
 @JsonAutoDetect(fieldVisibility = Visibility.PUBLIC_ONLY)
 trait NamingFactoryConfig {
 
-  def kind = getClass.getName
+  def kind = getClass.getCanonicalName
 
   /**
    * Construct a NamingFactory.